shared: move compareOwnerReferences from kubevirt
