+++ frontend/packages/kubevirt-plugin/src/selectors/pod/selectors.ts
@@ -1,5 +1,6 @@
 import { get } from 'lodash';
 import { getName, getNamespace, getOwnerReferences } from '@console/shared/src/selectors';
+import { compareOwnerReference } from '@console/shared/src/utils/owner-references';
 import { PodKind } from '@console/internal/module/k8s';
 import { getLabelValue } from '../selectors';
 import { VMKind } from '../../types';
@@ -9,7 +10,7 @@ import {
   STORAGE_IMPORT_PVC_NAME,
   VIRT_LAUNCHER_POD_PREFIX,
 } from '../../constants';
-import { compareOwnerReference, buildOwnerReferenceForModel } from '../../utils';
+import { buildOwnerReferenceForModel } from '../../utils';
 import { VirtualMachineInstanceModel } from '../../models';
 
 export const getHostName = (pod: PodKind) =>
