+++ frontend/packages/kubevirt-plugin/src/utils/index.ts
@@ -91,28 +91,6 @@ export const buildOwnerReferenceForModel = (
   uid,
 });
 
-export const compareOwnerReference = (
-  obj: OwnerReference,
-  otherObj: OwnerReference,
-  compareModelOnly?: boolean,
-) => {
-  if (obj === otherObj) {
-    return true;
-  }
-  if (!obj || !otherObj) {
-    return false;
-  }
-  const isUIDEqual = obj.uid && otherObj.uid ? compareModelOnly || obj.uid === otherObj.uid : true;
-  const isNameEqual = compareModelOnly || obj.name === otherObj.name;
-
-  return (
-    obj.apiVersion === otherObj.apiVersion &&
-    obj.kind === otherObj.kind &&
-    isNameEqual &&
-    isUIDEqual
-  );
-};
-
 export const isCPUEqual = (a: CPU, b: CPU) =>
   a.sockets === b.sockets && a.cores === b.cores && a.threads === b.threads;
